requirements-preview.txt
streamlit>=1.32,<2
Pillow>=10,<12
